+++ decidim-admin/app/controllers/decidim/admin/application_controller.rb
@@ -6,6 +6,7 @@ module Decidim
     class ApplicationController < ::DecidimController
       include NeedsOrganization
       include NeedsPermission
+      include Decidim::Admin::Concerns::NeedsAdminTermsAccepted
       include FormFactory
       include LocaleSwitcher
       include PayloadInfo