+++ decidim-admin/app/forms/decidim/admin/organization_form.rb
@@ -27,11 +27,14 @@ module Decidim
       translatable_attribute :welcome_notification_subject, String
       translatable_attribute :welcome_notification_body, String
 
+      translatable_attribute :admin_terms_of_use_body, String
+
       validates :welcome_notification_subject, :welcome_notification_body, translatable_presence: true, if: proc { |form| form.customize_welcome_notification }
 
       validates :name, presence: true
       validates :default_locale, :reference_prefix, presence: true
       validates :default_locale, inclusion: { in: :available_locales }
+      validates :admin_terms_of_use_body, translatable_presence: true
 
       private