+++ decidim-admin/spec/commands/decidim/admin/update_organization_spec.rb
@@ -15,6 +15,7 @@ module Decidim::Admin
             default_locale: "en",
             badges_enabled: true,
             user_groups_enabled: true,
+            admin_terms_of_use_body: Faker::Lorem.paragraph,
             send_welcome_notification: false
           }
         }