+++ decidim-core/lib/decidim/core/test/factories.rb
@@ -119,6 +119,7 @@ FactoryBot.define do
     about { "<script>alert(\"ABOUT\");</script>" + Faker::Lorem.paragraph(2) }
     confirmation_sent_at { Time.current }
     accepted_tos_version { organization.tos_version }
+    admin_terms_accepted_at { Time.current }
     email_on_notification { true }
 
     trait :confirmed do